Split sync diagnose action handlers
