+++ service-worker.js
@@ -157,6 +157,10 @@ const APP_SHELL = [
   '/js/sync-environment.js',
   '/js/sync-identity.js',
   '/js/sync-diagnostics.js',
+  '/js/sync-diagnose-actions-context.js',
+  '/js/sync-diagnose-relay-actions.js',
+  '/js/sync-diagnose-identity-actions.js',
+  '/js/sync-diagnose-cutover-actions.js',
   '/js/sync-diagnose-actions.js',
   '/js/sync-diagnose-ui.js',
   '/js/sync-actions.js',
